requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6

sqlalchemy==2.0.36
asyncpg==0.29.0
greenlet==3.0.3

redis==5.0.8

pydantic==2.10.3
python-dotenv==1.0.1

validators==0.22.0

slowapi==0.1.9

celery==5.4.0

aiosqlite==0.20.0
